requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.6.3=pypi_0
bottleneck=1.3.5=py38h7deecbd_0
brewer2mpl=1.4.1=py_3
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2024.2.2=py38h06a4308_0
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=unix_pyh707e725_0
cmake=3.29.2=pypi_0
colorcet=3.0.1=py38h06a4308_0
complexpytorch=0.4=pypi_0
configparser=7.0.0=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.2=py38h774aba0_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=py_0
docopt=0.6.2=pypi_0
einops=0.7.0=pyhd8ed1ab_1
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.9.0=py38h06a4308_0
fitlog=0.9.15=pypi_0
flask=3.0.0=pypi_0
fonttools=4.48.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.6.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.32=pyhd8ed1ab_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.57.0=pypi_0
huggingface-hub=0.22.2=pypi_0
icu=67.1=he1b5a44_0
idna=3.4=py38h06a4308_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.3.0=pyhd8ed1ab_1
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py38h6a678d5_0
kymatio=0.3.0=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.0.149=0
libcurand=10.3.3.53=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lightning-utilities=0.9.0=pyhd8ed1ab_0
lit=18.1.3=pypi_0
lz4-c=1.9.4=h6a678d5_0
mamba-ssm=1.2.0.post1=pypi_0
markdown=3.4.4=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mne=1.6.1=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py38h06a4308_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
ninja=1.11.1.1=pypi_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.24.3=py38h14f4228_0
numpy-base=1.24.3=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
packaging=22.0=py38h06a4308_0
pandas=1.4.4=py38h6a678d5_0
param=1.12.3=py38h06a4308_0
pathtools=0.1.2=py_1
pickle5=0.0.11=pypi_0
pillow=9.4.0=py38h6a678d5_0
pip=23.2.1=py38h06a4308_0
platformdirs=4.2.0=pypi_0
pooch=1.8.0=pypi_0
pot=0.9.3=pypi_0
protobuf=3.20.3=py38h6a678d5_0
psutil=5.9.0=py38h5eee18b_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyct=0.5.0=py38h06a4308_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyre-extensions=0.0.30=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py38h06a4308_0
pyyaml=5.1.2=py38h516909a_1
readline=7.0=h7b6447c_5
regex=2024.4.16=pypi_0
requests=2.31.0=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.2.2=py38h6a678d5_1
scipy=1.10.0=py38h14f4228_0
seaborn=0.12.2=py38h06a4308_0
sentry-sdk=1.29.0=pyhd8ed1ab_0
setproctitle=1.1.10=py38h516909a_1001
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.33.0=h62c20be_0
sympy=1.11.1=py38h06a4308_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
torch-complex=0.4.3=pyhd8ed1ab_0
torchaudio=2.0.0=py38_cu118
torcheval=0.0.6=pypi_0
torchmetrics=1.0.3=pyhd8ed1ab_0
torchtnt=0.2.0=pypi_0
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu118
tornado=6.1=py38h0a891b7_3
tqdm=4.66.1=pypi_0
transformers=4.40.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
urllib3=1.26.16=py38h06a4308_0
wandb=0.15.7=pyhd8ed1ab_0
werkzeug=3.0.0=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
